+++ src/keybinds/config.rs
@@ -113,6 +113,7 @@ pub enum Keybind {
 pub struct LogTabKeybindsConfig {
     pub mark_head: Option<Keybind>,
     pub goto_parent: Option<Keybind>,
+    pub goto_child: Option<Keybind>,
 
     pub duplicate: Option<Keybind>,
     pub create_new: Option<Keybind>,